netlist: improve readability. (nw)
Renamed cast member of the constants struct to magic to clearly identify
magic numbers.

Introduced nlconst struct inheriting from plib::constants<nl_fptype> to
make code better understandable.

+++ src/lib/netlist/nl_config.h
@@ -104,7 +104,7 @@ static constexpr const auto NETLIST_CLOCK = NETLIST_INTERNAL_RES;
 // due to numeric issues
 //============================================================
 
-using nl_fptype = double;
+using nl_fptype = float;
 
 namespace netlist
 {
